(improvement)(Headless) move Knowledge package to chat package
