+++ headless/core/src/main/java/com/tencent/supersonic/headless/core/chat/knowledge/DictWord.java
@@ -1,4 +1,4 @@
-package com.tencent.supersonic.headless.core.knowledge;
+package com.tencent.supersonic.headless.core.chat.knowledge;
 
 import java.util.Objects;
 import lombok.Data;
