+++ headless/core/src/main/java/com/tencent/supersonic/headless/core/chat/knowledge/EmbeddingResult.java
@@ -1,4 +1,4 @@
-package com.tencent.supersonic.headless.core.knowledge;
+package com.tencent.supersonic.headless.core.chat.knowledge;
 
 import com.google.common.base.Objects;
 import java.util.Map;
