+++ headless/core/src/main/java/com/tencent/supersonic/headless/core/chat/knowledge/builder/ModelWordBuilder.java
@@ -1,18 +1,20 @@
-package com.tencent.supersonic.headless.core.knowledge.builder;
-
+package com.tencent.supersonic.headless.core.chat.knowledge.builder;
 
 import com.google.common.collect.Lists;
 import com.tencent.supersonic.common.pojo.enums.DictWordType;
 import com.tencent.supersonic.headless.api.pojo.SchemaElement;
-import com.tencent.supersonic.headless.core.knowledge.DictWord;
+import com.tencent.supersonic.headless.core.chat.knowledge.DictWord;
 import java.util.List;
 import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+/**
+ * model word nature
+ */
 @Service
 @Slf4j
-public class EntityWordBuilder extends BaseWordWithAliasBuilder {
+public class ModelWordBuilder extends BaseWordWithAliasBuilder {
 
     @Override
     public List<DictWord> doGet(String word, SchemaElement schemaElement) {
@@ -23,16 +25,13 @@ public class EntityWordBuilder extends BaseWordWithAliasBuilder {
         result.add(getOneWordNature(word, schemaElement, false));
         result.addAll(getOneWordNatureAlias(schemaElement, false));
         return result;
-
     }
 
-    @Override
     public DictWord getOneWordNature(String word, SchemaElement schemaElement, boolean isSuffix) {
-        String nature = DictWordType.NATURE_SPILT + schemaElement.getModel()
-                + DictWordType.NATURE_SPILT + schemaElement.getId() + DictWordType.ENTITY.getTypeWithSpilt();
         DictWord dictWord = new DictWord();
         dictWord.setWord(word);
-        dictWord.setNatureWithFrequency(String.format("%s " + DEFAULT_FREQUENCY * 2, nature));
+        String nature = DictWordType.NATURE_SPILT + schemaElement.getDataSet();
+        dictWord.setNatureWithFrequency(String.format("%s " + DEFAULT_FREQUENCY, nature));
         return dictWord;
     }
