+++ headless/core/src/main/java/com/tencent/supersonic/headless/core/chat/knowledge/builder/TagWordBuilder.java
@@ -1,18 +1,16 @@
-package com.tencent.supersonic.headless.core.knowledge.builder;
+package com.tencent.supersonic.headless.core.chat.knowledge.builder;
 
 import com.google.common.collect.Lists;
 import com.tencent.supersonic.common.pojo.enums.DictWordType;
 import com.tencent.supersonic.headless.api.pojo.SchemaElement;
-import com.tencent.supersonic.headless.core.knowledge.DictWord;
+import com.tencent.supersonic.headless.core.chat.knowledge.DictWord;
+
 import java.util.List;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 
-/**
- * dimension word nature
- */
 @Service
-public class DimensionWordBuilder extends BaseWordWithAliasBuilder {
+public class TagWordBuilder extends BaseWordWithAliasBuilder {
 
     @Override
     public List<DictWord> doGet(String word, SchemaElement schemaElement) {
@@ -20,7 +18,7 @@ public class DimensionWordBuilder extends BaseWordWithAliasBuilder {
         result.add(getOneWordNature(word, schemaElement, false));
         result.addAll(getOneWordNatureAlias(schemaElement, false));
         String reverseWord = StringUtils.reverse(word);
-        if (StringUtils.isNotEmpty(word) && !word.equalsIgnoreCase(reverseWord)) {
+        if (!word.equalsIgnoreCase(reverseWord)) {
             result.add(getOneWordNature(reverseWord, schemaElement, true));
         }
         return result;
@@ -31,10 +29,10 @@ public class DimensionWordBuilder extends BaseWordWithAliasBuilder {
         dictWord.setWord(word);
         Long modelId = schemaElement.getModel();
         String nature = DictWordType.NATURE_SPILT + modelId + DictWordType.NATURE_SPILT + schemaElement.getId()
-                + DictWordType.DIMENSION.getTypeWithSpilt();
+                + DictWordType.TAG.getTypeWithSpilt();
         if (isSuffix) {
             nature = DictWordType.NATURE_SPILT + modelId + DictWordType.NATURE_SPILT + schemaElement.getId()
-                    + DictWordType.SUFFIX.getTypeWithSpilt() + DictWordType.DIMENSION.getTypeWithSpilt();
+                    + DictWordType.SUFFIX.getTypeWithSpilt() + DictWordType.TAG.getTypeWithSpilt();
         }
         dictWord.setNatureWithFrequency(String.format("%s " + DEFAULT_FREQUENCY, nature));
         return dictWord;
