+++ headless/core/src/main/java/com/tencent/supersonic/headless/core/chat/mapper/EmbeddingMatchStrategy.java
@@ -6,8 +6,8 @@ import com.tencent.supersonic.common.util.embedding.Retrieval;
 import com.tencent.supersonic.common.util.embedding.RetrieveQuery;
 import com.tencent.supersonic.common.util.embedding.RetrieveQueryResult;
 import com.tencent.supersonic.headless.core.config.OptimizationConfig;
-import com.tencent.supersonic.headless.core.knowledge.EmbeddingResult;
-import com.tencent.supersonic.headless.core.knowledge.MetaEmbeddingService;
+import com.tencent.supersonic.headless.core.chat.knowledge.EmbeddingResult;
+import com.tencent.supersonic.headless.core.chat.knowledge.MetaEmbeddingService;
 import com.tencent.supersonic.headless.core.pojo.QueryContext;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.CollectionUtils;
