+++ headless/core/src/main/java/com/tencent/supersonic/headless/core/chat/mapper/MapperHelper.java
@@ -3,7 +3,7 @@ package com.tencent.supersonic.headless.core.chat.mapper;
 import com.hankcs.hanlp.algorithm.EditDistance;
 import com.tencent.supersonic.headless.api.pojo.response.S2Term;
 import com.tencent.supersonic.headless.core.config.OptimizationConfig;
-import com.tencent.supersonic.headless.core.knowledge.helper.NatureHelper;
+import com.tencent.supersonic.headless.core.chat.knowledge.helper.NatureHelper;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
