+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/utils/DataSetSchemaBuilder.java
@@ -187,6 +187,7 @@ public class DataSetSchemaBuilder {
                     .alias(alias)
                     .relatedSchemaElements(getRelateSchemaElement(metric))
                     .defaultAgg(metric.getDefaultAgg())
+                    .dataFormatType(metric.getDataFormatType())
                     .build();
             metrics.add(metricToAdd);
